requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv

